requirements.txt
streamlit
keras
pillow
numpy
tensorflow